+++ CHANGELOG.md
@@ -45,6 +45,21 @@ Bug Fix:
 
 Breaking Change:
 - Remove Elysia Symbol (Internal)
+# 0.5.25 - 25 Jul 2023
+Bug fix:
+- ws resolve type to undefined instead of unknown cause unexpected type mismatched when not provided
+
+# 0.5.24 - 22 Jul 2023
+Bug fix:
+- [#68](https://github.com/elysiajs/elysia/issues/68) invalid path params when using numeric
+
+# 0.5.23 - 20 Jul 2023
+Bug fix:
+- [#68](https://github.com/elysiajs/elysia/issues/68) invalid optional query params when using numeric
+
+# 0.5.22 - 9 Jul 2023
+Bug fix:
+- update onAfterHandle to be Response only
 
 # 0.5.20 - 23 Jun 2023
 Bug fix: